thanh toan tien mua sach

+++ 4.7.c
@@ -10,7 +10,7 @@ main ()
   printf("ISBN: ?\n"); scanf("%7s",ISBN);
   printf("Title: ?\n"); scanf(" %[^\n]",Title);
   printf("Price= ?\n"); scanf("%8f",&Price);
-  VAT = (4 / 100);
+  VAT = (4 /(double) 100);
   Total = Price*Quantity;
   Pay = Total*(1 + VAT);
   printf("%s\n",frame);
